pyproject.toml
[tool.poetry]
name = "chat-app-arkleap"
version = "0.1.0"
description = "this is a repoitory for the interview at ArkLeap"
authors = ["basola21 <basel21mahmoud@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
